A new version of the server from the ground up.
I spent some time on linked in learning and actually learned the skills I needed to help build this instead of just reading documentation as I go. So here's a simple websocket server with Ping Pong mechanics to help maintain a connection and not have it drop due to a timeout.